backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
boto3==1.36.0
botocore>=1.36.0,<1.36.4
python-dotenv==1.0.0
pandas>=2.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
werkzeug==3.0.1
